pyproject.toml
[project]
name = "ai4eu_server_demo"
description = "A containerized demo application for AI on demand (previously AI4EU) "
version = "0.4.20230627"
requires-python = ">=3.11"
authors = [

    { name = "Antonis Ganios"},
    { name = "Íñigo Aréjula Aísa"},
    { name = "Jos van der Velde", email = "j.d.v.d.velde@tue.nl"},
    { name = "Pieter Gijsbers", email = "p.gijsbers@tue.nl"},
    { name = "Giannis Mouchakis"},
    {name = "Taniya Das", email = "t.das@tue.nl"}
]
dependencies = [
    "urllib3== 2.0.3",
    "bibtexparser==1.4.0",
    "datasets==2.13.1",
    "fastapi==0.98.0",
    "uvicorn==0.22.0",
    "requests==2.31.0",
    "mysqlclient==2.1.1",
    "oic==1.6.0",
    "python-keycloak==3.1.1",
    "python-dotenv==1.0.0",
    "pydantic_schemaorg==1.0.6",
    "python-dateutil==2.8.2",
    "sqlmodel==0.0.8",
    "httpx==0.24.1",
    "sickle==0.7.0",
    "xmltodict==0.13.0",
    "python-multipart==0.0.6"
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "types-python-dateutil==2.8.19.13",
    "pytest==7.4.0",
    "pytest-xdist==3.3.1",
    "pre-commit==3.3.3",
    "responses==0.23.1",
    "starlette==0.27.0"
]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::FutureWarning"
]
